pyproject.toml
[project]
name = "alt-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.2",
    "numpy>=1.24.0",
    "openai>=1.84.0",
    "opencv-python-headless>=4.9.0.80",
    "pillow>=10.3.0",
    "pymupdf==1.23.21",
    "requests>=2.31.0",
    "streamlit>=1.35.0",
    "tqdm>=4.66.0",
]
